Move filters definition to Text class.
Objectively thinking we append filters to text rather than some "filtering machine" which Filter is used to be.
Since this commit Filter is nothing more that consolidation of common code/properties for filters,
the way an abstract class is meant to be.
Before that Filter was a "filter manager" if called statically and a filter if it's an object.

Now Text takes management job and pass itself through filters chain when required.
It autoloads classes too.

You can still define global defailt filters with

    Text::defaultFilters()